refactor and changed default beam=1 and len_pen=0
